Replace NoopWorksheetProvider with Option[WorksheetProvider]
